finish eventCard.module.css and eventCard.tsx
